feat(tui): --snapshot-genes batch path
Resolves names against loaded AnnotationSources using the same
"union of matches per chrom" rule as AppState::find_gene_region;
skips unknowns with a warning. Reuses the BED batch render loop.

Also drops a redundant `view_width as u64` cast in region.rs (it
was already u64), silencing a clippy warning that pre-dated this
branch and would otherwise trip a clean clippy run on the new code.

+++ crates/igv-core/src/region.rs
@@ -128,7 +128,7 @@ pub fn genomic_to_screen(
     if view_width == 0 || screen_width == 0 {
         return None;
     }
-    if view_width as u64 > screen_width as u64 {
+    if view_width > screen_width as u64 {
         let scaled = (rel as u128 * screen_width as u128 / view_width as u128) as u32;
         Some(scaled.min(screen_width.saturating_sub(1)))
     } else {
@@ -143,7 +143,7 @@ pub fn screen_to_genomic(
     view_width: u64,
     screen_width: u32,
 ) -> u64 {
-    if view_width as u64 > screen_width as u64 {
+    if view_width > screen_width as u64 {
         let g = screen_pos as u128 * view_width as u128 / screen_width as u128;
         view_start + g as u64
     } else {
